requirements.txt
absl-py==1.4.0
addict==2.4.0
aiodns==3.1.1
aiohttp==3.8.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
appdirs==1.4.4
apturl==0.5.2
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
audio-recorder-streamlit==0.0.8
autocomplete==0.0.104
av==10.0.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.12.3
beniget==0.4.1
blinker==1.7.0
bottle==0.12.25
Brlapi==0.8.3
Brotli==1.0.9
cachetools==5.2.0
cchardet==2.1.7
certifi==2020.6.20
cffi==1.16.0
chardet==4.0.0
charset-normalizer==2.1.1
clarifai==10.0.1
clarifai-grpc==10.0.4
click==8.1.3
colorama==0.4.4
coloredlogs==15.0.1
comm==0.1.3
command-not-found==0.3
commonmark==0.9.1
contextlib2==21.6.0
contourpy==1.2.0
cryptography==3.4.8
cuda-python==12.2.0
cupshelpers==1.0
cvzone==1.6.1
cycler==0.12.1
Cython==3.0.2
dataclasses-json==0.6.3
dbus-python==1.2.18
debugpy==1.6.7
decorator==4.4.2
decord==0.4.1
defer==1.0.6
Deprecated==1.2.14
distro==1.7.0
distro-info==1.1+ubuntu0.1
duplicity==0.8.21
einops==0.6.1
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==1.2.0
fasteners==0.14.1
filelock==3.9.0
Flask==3.0.0
flatbuffers==23.5.26
fonttools==4.47.2
frozenlist==1.4.1
fs==2.4.12
fsspec==2023.12.2
ftfy==6.1.1
future==0.18.2
gast==0.5.4
gdown==5.1.0
gitdb==4.0.11
GitPython==3.1.41
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.62.0
grad-cam==1.5.0
graphviz==0.20.1
greenlet==3.0.3
grpcio==1.58.0
h11==0.14.0
h5py==3.8.0
hiddenlayer==0.3
html5lib==1.1
httpcore==1.0.2
httplib2==0.20.2
httpx==0.26.0
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.3
imagecodecs==2024.1.1
imageio==2.28.1
imgaug==0.4.0
imgviz==1.7.5
importlib-metadata==6.8.0
imutils==0.5.3
ipykernel==6.23.0
ipython==8.13.2
itsdangerous==2.1.2
jedi==0.18.2
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.2.0
jupyter_core==5.3.0
kdu-jp2==1.2
keras==2.14.0
keras-layer-normalization==0.16.0
keyring==23.5.0
kiwisolver==1.4.5
labelme==5.4.1
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.2
libclang==16.0.6
lightning-utilities==0.10.0
llama-index==0.9.34
llvmlite==0.38.0
lmdb==1.4.1
lockfile==0.12.2
louis==3.20.0
lpips==0.1.4
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.2
matlabengine==23.2.1
matplotlib==3.5.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mediapipe==0.10.9
memory-profiler==0.61.0
ml-dtypes==0.2.0
mlp-mixer-pytorch==0.1.1
mmaction2==1.1.0
mmcv==1.3.11
mmengine==0.0.1rc0
monotonic==1.6
more-itertools==8.10.0
mpmath==0.0.0
multidict==6.0.4
mypy-extensions==1.0.0
natsort==8.4.0
nest-asyncio==1.5.9
netifaces==0.11.0
netron==7.3.9
networkx==3.1
nltk==3.8.1
numba==0.55.1
numpy==1.26.3
oauthlib==3.2.0
olefile==0.46
onnxruntime==1.17.1
onx==0.3.1
openai==1.8.0
openai-clip==1.0.1
opencv-contrib-python==4.8.0.76
opencv-python==4.9.0.80
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.2
pandas==1.5.3
paramiko==2.9.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.2.0
platformdirs==3.5.1
ply==3.11
progressbar33==2.4
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==14.0.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycairo==1.20.1
pycares==4.4.0
pycparser==2.21
pycups==2.0.1
pydantic==1.9.2
pydantic_core==2.14.6
pydeck==0.8.1b0
pydub==0.25.1
pyenchant==3.2.2
pyfiglet==0.8.post1
pygame==2.5.2
Pygments==2.17.2
PyGObject==3.42.1
pyiqa==0.1.7
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
PyOpenGL==3.1.7
pyparsing==3.1.1
pypdf==4.0.0
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyRFC3339==1.1
PySocks==1.7.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.0
python-rapidjson==1.14
pythran==0.10.0
pytorch-lightning==2.1.3
pytorch-msssim==1.0.0
pytorch-triton==2.1.0+e6216047b8
pyttsx3==2.90
pytz==2022.1
PyWavelets==1.4.1
pyxdg==0.27
PyYAML==6.0.1
pyzmq==25.0.2
QtPy==2.4.1
referencing==0.32.1
regex==2023.8.8
reportlab==3.6.8
requests==2.31.0
requests-oauthlib==1.3.1
rich==12.6.0
rpds-py==0.17.1
rsa==4.9
safetensors==0.3.3
schema==0.7.5
scikit-image==0.20.0
scikit-learn==1.3.0
scipy==1.11.2
screen-resolution-extra==0.0.0
seaborn==0.13.2
SecretStorage==3.3.1
shapely==2.0.1
single-source==0.3.0
six==1.16.0
sklearn==0.0.post9
smmap==5.0.1
sniffio==1.3.0
sounddevice==0.4.6
soupsieve==2.3.1
SQLAlchemy==2.0.25
ssh-import-id==5.11
SSIM-PIL==1.0.14
stack-data==0.6.2
streamlit==1.30.0
sympy==1.9
systemd-python==234
tb-nightly==2.15.0a20230907
tenacity==8.2.3
tensorboard==2.14.1
tensorboard-data-server==0.7.1
tensorboardX==2.6.2.2
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
textual==0.1.18
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tifffile==2023.4.12
tiktoken==0.5.2
timm==0.9.7
tk==0.1.0
tokenizers==0.14.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.2.0.dev20230907+cu121
torch-cka==0.21
torchaudio==2.2.0.dev20230907+cu121
torchmetrics==1.2.1
torchsummary==1.5.1
torchview==0.2.6
torchvision==0.17.0.dev20230907+cu121
torchviz==0.0.2
tornado==6.3.2
tqdm==4.66.1
traitlets==5.9.0
transformers==4.35.0
tritonclient==2.41.1
ttach==0.0.3
typing-inspect==0.9.0
typing_extensions==4.9.0
tzlocal==5.2
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufoLib2==0.13.1
ufw==0.36.1
ultralytics==8.2.2
unattended-upgrades==0.1
unicodedata2==14.0.0
urllib3==1.26.5
usb-creator==0.3.7
validators==0.22.0
vision-transformer-pytorch==1.0.3
wadllib==1.3.6
watchdog==3.0.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
Werkzeug==3.0.1
wrapt==1.14.1
xdg==5
xkit==0.0.0
xlrd==2.0.1
yapf==0.40.1
yarl==1.9.4
zipp==1.0.0
